+++ src/Pfam/Pfam.jl
@@ -4,11 +4,14 @@ module Pfam
   using MIToS.MSA
   using MIToS.SIFTS
   using MIToS.PDB
+  using MIToS.Information
   using PairwiseListMatrices
   using DataStructures
+  using ROCAnalysis
 
   export  Stockholm, downloadpfam, getseq2pdb,
-          msacolumn2pdbresidue, msacontacts
+          msacolumn2pdbresidue, msacontacts,
+          get_contact_mask, AUC
 
   include("download.jl")
   include("pdb.jl")