+++ test/msaannotations.jl
@@ -38,15 +38,16 @@ let pfam = read(joinpath(pwd(), "data", "PF09645_full.stockholm"), Stockholm,
   """)
   print(io, pfam.annotations)
   printed = split(takebuf_string(io), '\n')
-  @test length(printed) == 16
-  @test printed[1] == string("#=GF ColMap   6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22",
+  @test length(printed) == 17
+  @test printed[1] == string("#=GF NCol   120")
+  @test printed[2] == string("#=GF ColMap   6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22",
                              ",23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,",
                              "43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,",
                              "63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,",
                              "84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101,102,",
                              "103,104,105,106,107,108,109,110,111,112,113,114,115")
-  @test ismatch(r"MIToS_", printed[2])
   @test ismatch(r"MIToS_", printed[3])
+  @test ismatch(r"MIToS_", printed[4])
 
   print("""
   Test delete MIToS annotations
@@ -54,7 +55,7 @@ let pfam = read(joinpath(pwd(), "data", "PF09645_full.stockholm"), Stockholm,
   delete_annotated_modifications!(pfam)
   print(io, pfam.annotations)
   printed = split(takebuf_string(io), '\n')
-  @test length(printed) == 16 - 2 # 2 MIToS annotations
-  @test !ismatch(r"MIToS_", printed[2])
+  @test length(printed) == 17 - 2 # 2 MIToS annotations
   @test !ismatch(r"MIToS_", printed[3])
+  @test !ismatch(r"MIToS_", printed[4])
 end